backup/requirements.txt
psycopg2-binary==2.9.10
schedule==1.2.2
dotenv==0.9.9